Rename to pathtable
