--- internal/bubble-lookup/config.go
@@ -1,6 +1,0 @@
-package bubblelookup
-
-// Config is used to configure a bubble-lookup Client
-type Config struct {
-	BaseDirectory string
-}
