+++ internal/pathtable/README.md
