+++ internal/pathtable/client.go
@@ -1,6 +1,6 @@
-package bubblelookup
+package pathtable
 
-// Client is the interface for bubble-lookup methods
+// Client is the interface for pathtable methods
 type Client interface {
 	Get(path string) ([]AliasInfo, error)
 	Set(path string, info AliasInfo) error
@@ -10,7 +10,7 @@ type client struct {
 	baseDirectory string
 }
 
-// NewClient instantiates a bubble-lookup client
+// NewClient instantiates a pathtable client
 func NewClient(config *Config) Client {
 	return &client{
 		baseDirectory: config.BaseDirectory,
