+++ internal/pathtable/config.go
@@ -1,0 +1,6 @@
+package pathtable
+
+// Config is used to configure a pathtable Client
+type Config struct {
+	BaseDirectory string
+}
