+++ internal/pathtable/set_test.go
@@ -1,4 +1,4 @@
-package bubblelookup
+package pathtable
 
 import (
 	"testing"
