+++ internal/pathtable/test_utils.go
@@ -1,4 +1,4 @@
-package bubblelookup
+package pathtable
 
 import (
 	"os"
